Major Update

+++ programs/LanguageModel.py
@@ -3,11 +3,13 @@ import subprocess
 import Ngram
 import DownloadFiles
 import CleanData
+import JoinReduce
 import Formats as of
 import pandas as pd
 from shutil import rmtree
 import os
 import numpy as np
+import gc
 
 
 class GenerateLanguageModel:
@@ -16,9 +18,18 @@ class GenerateLanguageModel:
         self.color01 = "\033[92m"  # Green
 
         # Create log data ----------------------------------------------------------------------------------------------
+        LogVars = ["ProcessID", "Date", "Start_LanguageModel", "Time_Total", "Version", "PC", "Method", "Seed",
+                   "SampleTrainRate", "Count_SampleTrainLines", "TestModel", "SampleTestRate", "Time_LanguageModel",
+                   "Time_DownloadScript", "Time_Download", "Time_WriteCorpus", "FileSize_Corpus", "Count_CorpusLines",
+                   "Time_CleanScript", "Time_ReadCorpus", "FileSize_TrainTokens", "Count_WordTokens", "Time_cleanCorpus", "Count_TokensLines",
+                   "Start_Ngram", "Time_CreateNGram", "Time_NGramData", "Stop_Ngram"]
+
         self.infoLog = pd.DataFrame(np.array([[ID]]), columns=['ProcessID'])
+        self.infoLog = self.infoLog.reindex(columns = LogVars)
+        self.infoLog['Date'] = of.getDate()
         self.infoLog['Version'] = Version
         self.infoLog['Method'] = Method
+        self.infoLog['PC'] = "ASUS"
         self.infoLog['Start_LanguageModel'] = of.calcTime()
         StartScript = of.calcTime()
         self.infoLog['SampleTrainRate'] = SampleTrainRate
@@ -27,7 +38,8 @@ class GenerateLanguageModel:
         self.infoLog['Seed'] = Seed
 
         # Print Start --------------------------------------------------------------------------------------------------
-        of.StartModel(phrase=str("Creating Language Model - Version: " + str(Version) + "; Method: " + str(Method)),
+        of.StartModel(phrase=str("Creating Language Model - Version: " + str(Version) + "; Method: " + str(Method) +
+                                 "; Sample Rate: " + str(SampleTrainRate)),
                       time=StartScript, k=140)
 
         # Download Files -----------------------------------------------------------------------------------------------
@@ -37,31 +49,31 @@ class GenerateLanguageModel:
         CleanData.CleanCorpus(Corpus="../data/Corpus.txt", infoLog = self.infoLog)
 
         # Create Sample ------------------------------------------------------------------------------------------------
-        if os.path.isfile("../data/TrainTokens.txt") != 1:
-            startTime = of.calcTime()
-            of.ElapseStart(startTime, "Reading Tokens Data")
-            with open(File, "r") as file:
-                Data = file.readlines()
-            N = len(Data)
-            n = int(round(N*SampleTrainRate, 0))
-            self.infoLog['Count_TokensLines'] = N
-            self.infoLog['Count_SampleTrainLines'] = n  # Sample Size
-            of.ElapseEnd(startTime)
+        # if os.path.isfile("../data/TrainTokens.txt") != 1:
+        startTime = of.calcTime()
+        of.ElapseStart(startTime, "Reading Tokens Data")
+        with open(File, "r") as file:
+            Data = file.readlines()
+        N = len(Data)
+        n = int(round(N*SampleTrainRate, 0))
+        self.infoLog['Count_TokensLines'] = N
+        self.infoLog['Count_SampleTrainLines'] = n  # Sample Size
+        of.ElapseEnd(startTime)
 
-            # Creating Train Data
-            startTime = of.calcTime()
-            of.ElapseStart(startTime, "Sampling Tokens Data")
-            self.TrainData = self.sampling(Data=Data, output="TrainTokens.txt", N=N, Size=n, seed=Seed, Type="train")
-            self.infoLog['FileSize_TrainTokens'] = round(os.path.getsize("../data/TrainTokens.txt") / (1024 * 1024.0),
-                                                         2)
-            of.ElapseEnd(startTime)
+        # Creating Train Data
+        startTime = of.calcTime()
+        of.ElapseStart(startTime, "Sampling Tokens Data")
+        self.TrainData = self.sampling(Data=Data, output="TrainTokens.txt", N=N, Size=n, seed=Seed, Type="train")
+        FileSize_TrainTokens = round(os.path.getsize("../data/TrainTokens.txt") / (1024 * 1024.0), 2)
+        self.infoLog['FileSize_TrainTokens'] = FileSize_TrainTokens
+        of.ElapseEnd(startTime)
 
         # Create Model -------------------------------------------------------------------------------------------------
         self.createModel(Method=Method)
 
         # Print Finish Log ---------------------------------------------------------------------------------------------
         self.infoLog['Stop_Ngram'] = of.calcTime()
-        self.infoLog['Time_LanguageModel'] = of.evalElapse(start=StartScript)
+        self.infoLog['Time_LanguageModel'] = of.deltaTime(start=StartScript)
         of.EndScript(start=StartScript, phrase="Language Model Created")
 
         # TODO: Test Sample --------------------------------------------------------------------------------------------
@@ -110,10 +122,16 @@ class GenerateLanguageModel:
 
         # Save Log -----------------------------------------------------------------------------------------------------
         print(self.color01 + "\n>>> " + of.calcTime() + "\033[0m" + " Saving Log")
+        self.infoLog['Time_Total'] = float(of.deltaTime(StartScript))
         # self.infoLog = pd.DataFrame([self.infoLog])
         filename = "../data/Log.csv"
-        self.infoLog.to_csv(filename, index=False, encoding='utf-8')
-        print(self.infoLog)
+        if os.path.isfile(filename) == 1:
+            tempLog = pd.read_csv(filename)
+            LogFile = tempLog.append(self.infoLog)
+            LogFile.to_csv(filename, index=False, encoding='utf-8')
+        else:
+            self.infoLog.to_csv(filename, index=False, encoding='utf-8')
+        # print(self.infoLog)
         print(self.color01 + ">>> " + of.calcTime() + "\033[0m" + " Process Finished")
 
     # Sampling Function ------------------------------------------------------------------------------------------------
@@ -130,23 +148,35 @@ class GenerateLanguageModel:
 
     # Create Model Function --------------------------------------------------------------------------------------------
     def createModel(self, Method):
-        self.infoLog['Start_Ngram'] = of.calcTime()
+        Start_Ngram = of.calcTime()
+        self.infoLog['Start_Ngram'] = Start_Ngram
+
         # Creating Language Model according with the method indicated
         if Method.lower() == "mapreduce":
-            of.StartScript(self.infoLog['Start_Ngram'], "Executing MapReduce Method")
+            of.StartScript(Start_Ngram, "Executing MapReduce Method")
             # Executing shell script.
             # TODO: Read Data from argument instead of hardcoded
             output = "../data/MapReduce"
             mapReduce = "'python Mapper.py' 'python Reducer.py'"
-            blocksize = "15m"
+            blocksize = "5m"
             reducers = "8"
-            shell_command = ["cat 'TrainTokens.txt' | ./mc-hdfs.sh " + blocksize + " " +
+            shell_command = ["cat '../data/TrainTokens.txt' | ./mc-hdfs.sh " + str(blocksize) + " " +
                              reducers + " " + mapReduce + " " + output]
             if os.path.exists(output):
                 rmtree(output)
+                Start_Time = of.calcTime()
                 subprocess.check_call(shell_command, shell=True)
+                self.infoLog['Time_CreateNGram'] = float(of.deltaTime(Start_Time))
+                Start_Time = of.calcTime()
+                JoinReduce.JoinReduceFiles(DataPath='../data/', infoLog=self.infoLog)
+                self.infoLog['Time_NGramData'] = float(of.deltaTime(Start_Time))
             else:
+                Start_Time = of.calcTime()
                 subprocess.check_call(shell_command, shell=True)
+                self.infoLog['Time_CreateNGram'] = float(of.deltaTime(Start_Time))
+                Start_Time = of.calcTime()
+                JoinReduce.JoinReduceFiles(DataPath='../data/', infoLog=self.infoLog)
+                self.infoLog['Time_NGramData'] = float(of.deltaTime(Start_Time))
 
         elif Method.lower() == "sequential":
             of.StartScript(self.infoLog['Start_Ngram'], "Executing Sequential Method")
@@ -156,7 +186,9 @@ class GenerateLanguageModel:
 
         else:
             of.NormalMessage(" is not an available method.")
+            self.infoLog['Time_Ngram'] = None
 
-
-GenerateLanguageModel(ID="0001", File="../data/Tokens.txt", Method="sequential", Version="Git 0.1",
-                      SampleTrainRate=0.05, TestModel="No", SampleTestRate=0.01, Seed=17895)
+for i in range(1, 10):
+    GenerateLanguageModel(ID=str(i), File="../data/Tokens.txt", Method="sequential", Version="Git 0.1",
+                          SampleTrainRate=round(i/100, 2), TestModel="No", SampleTestRate=0.01, Seed=17895)
+    gc.collect()
